boil down to using exec
